requirements.txt
flask
nltk
pyyaml